Import reads directly from ftp/http site that user specify
Hi Dan, 
Uploading of read files directly from ftp/http site is missing there. It is very important to have capability to import read files from a ftp or http url as the read files are in giga bytes and it will help users to directly import the reads from site. Thanks.
Sunita