Update [WIP] RNA-seq - New Tools, 2016 Q3
Since we are increasingly using PDs as planning documents in addition to product specs, I have updated it to only the tools we plan to complete in two sprints allocated to this effort and if time permits add new ones. I have still kept the visualization tools (with a note that it is not guaranteed to be delivered in Q3).

FastQC and other read mgmt. tools, would be covered in another PD.