Trivial changes (other than changing name)
I don't see any way in GitHub to post a comment (not an Issue) about a file, other than by making a PR. Then other can comment on my comment.

I was wondering if we should add an optional section to PDs to list dependencies and/or risks. Some PDs have those in the Description, but it might be helpful to have it pulled out into a specific section.